refactor: Return uint32_t from GetSerializeSize, cast for weight calculations
Use uint32_t for serialization sizes since they fit in 32 bits.
Promote to uint64_t only when multiplying by WITNESS_SCALE_FACTOR to prevent overflow.

+++ src/flatfile.cpp
@@ -10,7 +10,7 @@
 #include <tinyformat.h>
 #include <util/fs_helpers.h>
 
-FlatFileSeq::FlatFileSeq(fs::path dir, const char* prefix, size_t chunk_size) :
+FlatFileSeq::FlatFileSeq(fs::path dir, const char* prefix, uint32_t chunk_size) :
     m_dir(std::move(dir)),
     m_prefix(prefix),
     m_chunk_size(chunk_size)
@@ -54,16 +54,16 @@ FILE* FlatFileSeq::Open(const FlatFilePos& pos, bool read_only) const
     return file;
 }
 
-size_t FlatFileSeq::Allocate(const FlatFilePos& pos, size_t add_size, bool& out_of_space) const
+uint64_t FlatFileSeq::Allocate(const FlatFilePos& pos, uint64_t add_size, bool& out_of_space) const
 {
     out_of_space = false;
 
-    unsigned int n_old_chunks = (pos.nPos + m_chunk_size - 1) / m_chunk_size;
-    unsigned int n_new_chunks = (pos.nPos + add_size + m_chunk_size - 1) / m_chunk_size;
+    uint64_t n_old_chunks = (pos.nPos + m_chunk_size - 1) / m_chunk_size;
+    uint64_t n_new_chunks = (pos.nPos + add_size + m_chunk_size - 1) / m_chunk_size;
     if (n_new_chunks > n_old_chunks) {
-        size_t old_size = pos.nPos;
-        size_t new_size = n_new_chunks * m_chunk_size;
-        size_t inc_size = new_size - old_size;
+        uint64_t old_size = pos.nPos;
+        uint64_t new_size = n_new_chunks * m_chunk_size;
+        uint64_t inc_size = new_size - old_size;
 
         if (CheckDiskSpace(m_dir, inc_size)) {
             FILE *file = Open(pos);
