+++ src/flatfile.h
@@ -47,7 +47,7 @@ class FlatFileSeq
 private:
     const fs::path m_dir;
     const char* const m_prefix;
-    const size_t m_chunk_size;
+    const uint32_t m_chunk_size;
 
 public:
     /**
@@ -57,7 +57,7 @@ public:
      * @param prefix A short prefix given to all file names.
      * @param chunk_size Disk space is pre-allocated in multiples of this amount.
      */
-    FlatFileSeq(fs::path dir, const char* prefix, size_t chunk_size);
+    FlatFileSeq(fs::path dir, const char* prefix, uint32_t chunk_size);
 
     /** Get the name of the file at the given position. */
     fs::path FileName(const FlatFilePos& pos) const;
@@ -74,7 +74,7 @@ public:
      * @param[out] out_of_space Whether the allocation failed due to insufficient disk space.
      * @return The number of bytes successfully allocated.
      */
-    size_t Allocate(const FlatFilePos& pos, size_t add_size, bool& out_of_space) const;
+    uint64_t Allocate(const FlatFilePos& pos, uint64_t add_size, bool& out_of_space) const;
 
     /**
      * Commit a file to disk, and optionally truncate off extra pre-allocated bytes if final.
